organize files & dirs (mostly backend structure)
