+++ src/backends/plonky2/mod.rs
@@ -1,6 +1,4 @@
 pub mod basetypes;
-pub mod common;
-pub mod main;
-pub mod mock_main;
-pub mod mock_signed;
+pub mod circuits;
+pub mod mock;
 pub mod primitives;
